reconcide log's wplit_writer implementations
close #1311
