+++ kube/client/client.go
@@ -105,7 +105,9 @@ func (kc *KubeClient) GetLogs(ctx context.Context, projectName string, consumer 
 	for _, pod := range pods.Items {
 		request := kc.client.CoreV1().Pods(kc.namespace).GetLogs(pod.Name, &corev1.PodLogOptions{Follow: follow})
 		service := pod.Labels[compose.ServiceTag]
-		w := utils.GetWriter(service, pod.Name, consumer)
+		w := utils.GetWriter(pod.Name, service, string(pod.UID), func(event compose.ContainerEvent) {
+			consumer.Log(event.Name, event.Source, event.Line)
+		})
 
 		eg.Go(func() error {
 			r, err := request.Stream(ctx)
